% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force method with option}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{must be a data frame with variables v (value) and w (weight)}

\item{W}{integer. Specify the maximum value for the knapsack.}

\item{parallel}{logical. Default is TRUE}
}
\value{
Returns a list of the maximum value and which elemets used.
}
\description{
Calulate with the brute force method how to get the maximum value out of optimal weight. A choice of a paralell comutation is possible.
}
\examples{
knapsack_objects <-
data.frame(
 w=sample(1:4000, size = 2000, replace = TRUE),
 v=runif(n = 2000, 0, 10000))
 brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500,parallel = FALSE)
}
