% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{must be a data frame with variables v (value) and w (weight)}

\item{W}{integer. Specify the maximum value for the knapsack.}
}
\value{
Returns a list of the maximum value and which elemets used.
}
\description{
Calulate with the greedy method how to get the maximum value out of optimal weight.
}
\examples{
knapsack_objects <-
data.frame(
 w=sample(1:4000, size = 2000, replace = TRUE),
 v=runif(n = 2000, 0, 10000))
 greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
 greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)
}
