% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6VarmKorv.R
\docType{package}
\name{Lab6VarmKorv}
\alias{Lab6VarmKorv}
\alias{Lab6VarmKorv-package}
\title{Lab6VarmKorv: Calculate the knapsack problem with different programing styles!}
\description{
This package takes a data frame with different weights and values and locates the maximum. This is done by using a brute force method and an option to parallel process it. A dynamic programming method and a greedy method is implemented.
}
\section{The packages functions}{

\describe{
  \item{data_address}{Data.frame of longitude and latitude for the address/destination}
   \item{plot_address}{Plot the address/destination}
}
}

\author{
Albin Vasterlund and Eric Herwin
}
