% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_brute_force.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{Longitude and Latitude of a address/destination}
\usage{
knapsack_brute_force(x, W, parallel = FALSE)
}
\arguments{
\item{x}{must be a data frame with variables v (value) and w (weight)}

\item{W}{integer. Specify the maximum value for the knapsack.}

\item{parallel}{logical. Default is TRUE}
}
\value{
Returns a list of the maximum value and which elemets used.
}
\description{
Return the longitude and latitude of a address/destination
}
\examples{
knapsack_objects <-
data.frame(
 w=sample(1:4000, size = 2000, replace = TRUE),
 v=runif(n = n, 0, 10000))
 knapsack_brute_force(x = knapsack_objects[1:8,], W = 3500,parallel = FALSE)
 knapsack_brute_force(x = knapsack_objects[1:8,], W = 3500,parallel = TRUE)
}
